cli/requirements.txt
rich>=13.0
click>=8.1
psycopg2-binary>=2.9
httpx>=0.27
python-dotenv>=1.0
